requirements.txt
python-telegram-bot
pymongo
celery
redis
